Serializer: refactoring.
 - Returned back `Serializer` instantiation using `new` function.
 - Added `with` function to instantiate `Serializer` with specified
   struct encoder.
